requirements.txt
aiogram==3.4.1
fastapi
uvicorn
sqlalchemy
psycopg[binary]
python-dotenv
cryptography